arreglado commit push reject -se agregaron las tablas listado de veterinarios y gestores para rol gestor -optimizado el filtro de usuarios por rol añadiendo una nueva columba a la entidad Usuario -arreglado que se volviera a crear el arbol de Usuarios cada vez que se daba click en el lobby de cada rol

+++ src/main/java/com/unal/una_huella/UNaHuellaLauncher/Entities/Role.java
@@ -12,7 +12,7 @@ public class Role implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO, generator = "native")
     @GenericGenerator(name = "native", strategy = "native")
-    private Long id;
+    private int id;
 
     @Column
     private String name;
@@ -20,11 +20,11 @@ public class Role implements Serializable {
     @Column
     private String description;
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id_role) {
+    public void setId(int id_role) {
         this.id = id_role;
     }
